requirements-inference.txt
torch
joblib
pandas
numpy
paho-mqtt
torchmetrics
lightning
scikit-learn
matplotlib
seaborn
wandb